+++ src/entities/cms_partials.rs
@@ -6,8 +6,8 @@ use sea_orm::entity::prelude::*;
 #[sea_orm(table_name = "cms_partials")]
 pub struct Model {
   #[sea_orm(primary_key)]
-  pub id: i32,
-  pub parent_id: Option<i32>,
+  pub id: i64,
+  pub parent_id: Option<i64>,
   pub name: String,
   #[sea_orm(column_type = "Text", nullable)]
   pub content: Option<String>,
